tried adding split

+++ src/libc/string.h
@@ -8,7 +8,7 @@ void int_to_ascii(int n, char str[]);
 void hex_to_ascii(int n, char str[]);
 
 void reverse(char s[]);
-int strlen(char s[]);
+size_t strlen(char s[]);
 void backspace(char s[]);
 void backspaces(char s[], int times);
 void append(char s[], char n);
@@ -16,7 +16,7 @@ int strcmp(char s1[], char s2[]);
 char* get_first_split(char* str, char split);
 void merge_strings(char str1[], char str2[], char result[]);
 bool has_char(char* str, char target);
-int split(char str[], char splitter, char* splitten[]);
+size_t split(char str[], char splitter, char* splitten[]);
 int count(char str[], char target);
 
 void lower(char* str);
